requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.1
openai==0.28.1
pymongo==4.5.0